chore: rename integration tests to e2e
The tests in integration/ deploy the full CDK stack to real AWS,
exercise the complete Step Functions workflow, and assert on live
DynamoDB state — that's E2E, not integration. Rename the directory,
build tag, CI job, and Makefile target accordingly.

Closes #55
