+++ e2e/integration_test.go
@@ -1,6 +1,6 @@
-//go:build integration
+//go:build e2e
 
-package integration
+package e2e
 
 import (
 	"os"
